+++ example_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/mozillazg/request"
 )
 
-func ExampleGet() {
+func ExampleRequest_Get() {
 	c := new(http.Client)
 	req := request.NewRequest(c)
 	url := "http://httpbin.org/get"
@@ -23,7 +23,21 @@ func ExampleGet() {
 	//http://httpbin.org/get
 }
 
-func ExampleGet_params() {
+func ExampleGet() {
+	c := new(http.Client)
+	args := request.NewArgs(c)
+	url := "http://httpbin.org/get"
+	resp, _ := request.Get(url, args)
+	d, _ := resp.Json()
+	defer resp.Body.Close()
+	fmt.Println(resp.Ok())
+	fmt.Println(d.Get("url").MustString())
+	// Output:
+	//true
+	//http://httpbin.org/get
+}
+
+func ExampleRequest_Get_params() {
 	c := new(http.Client)
 	req := request.NewRequest(c)
 	req.Params = map[string]string{
@@ -39,7 +53,7 @@ func ExampleGet_params() {
 	//http://httpbin.org/get?a=1&b=2
 }
 
-func ExampleGet_customHeaders() {
+func ExampleRequest_Get_customHeaders() {
 	c := new(http.Client)
 	req := request.NewRequest(c)
 	req.Headers = map[string]string{
@@ -57,7 +71,7 @@ func ExampleGet_customHeaders() {
 	//abc
 }
 
-func ExamplePost() {
+func ExampleRequest_Post() {
 	c := new(http.Client)
 	req := request.NewRequest(c)
 	req.Data = map[string]string{
@@ -69,7 +83,19 @@ func ExamplePost() {
 	defer resp.Body.Close()
 }
 
-func ExampleGet_cookies() {
+func ExamplePost() {
+	c := new(http.Client)
+	args := request.NewArgs(c)
+	args.Data = map[string]string{
+		"a": "1",
+		"b": "2",
+	}
+	url := "http://httpbin.org/post"
+	resp, _ := request.Post(url, args)
+	defer resp.Body.Close()
+}
+
+func ExampleRequest_Get_cookies() {
 	c := new(http.Client)
 	req := request.NewRequest(c)
 	req.Cookies = map[string]string{
@@ -81,20 +107,20 @@ func ExampleGet_cookies() {
 	defer resp.Body.Close()
 }
 
-func ExamplePost_files() {
+func ExampleRequest_Post_files() {
 	c := new(http.Client)
 	req := request.NewRequest(c)
 	f, _ := os.Open("test.txt")
 	defer f.Close()
 	req.Files = []request.FileField{
-		request.FileField{"abc", "abc.txt", f},
+		{"abc", "abc.txt", f},
 	}
 	url := "http://httpbin.org/post"
 	resp, _ := req.Post(url)
 	defer resp.Body.Close()
 }
 
-func ExamplePostRawBody() {
+func ExampleRequest_Post_rawBody() {
 	c := new(http.Client)
 	req := request.NewRequest(c)
 	req.Body = strings.NewReader("a=1&b=2&foo=bar")