+++ headers.go
@@ -2,7 +2,10 @@ package request
 
 import "net/http"
 
+// DefaultUserAgent define default User-Agent header
 var DefaultUserAgent = "go-request/" + Version
+
+// DefaultHeaders define default headers
 var DefaultHeaders = map[string]string{
 	"Connection":      "keep-alive",
 	"Accept-Encoding": "gzip, deflate",
@@ -10,10 +13,10 @@ var DefaultHeaders = map[string]string{
 	"User-Agent":      DefaultUserAgent,
 }
 
-// Default Content-Type Header for form body
+// DefaultContentType define default Content-Type Header for form body
 var DefaultContentType = "application/x-www-form-urlencoded; charset=utf-8"
 
-// Default Content-Type Header for json body
+// DefaultJsonType define default Content-Type Header for json body
 var DefaultJsonType = "application/json; charset=utf-8"
 
 func applyHeaders(a *Args, req *http.Request, contentType string) {